ProductService refactoring
